+++ newton/_src/solvers/mujoco/solver_mujoco.py
@@ -4019,6 +4019,8 @@ class SolverMuJoCo(SolverBase, CouplingInterface):
                 contacts.rigid_contact_damping,
                 contacts.rigid_contact_friction,
                 model.shape_margin,
+                model.shape_material_kf,
+                self.mjw_model.opt.impratio_invsqrt,
                 bodies_per_world,
                 self.newton_shape_to_mjc_geom,
                 # Mujoco warp contacts